made folders in c_cpp to organize it and added code for bst with commetns.
